+++ WPFTesting/WPFTesting/ViewModel/AnalizorModel.cs
@@ -123,39 +123,9 @@ public class AnalizorModel
 	}
 	public void AdvanceTime()
 	{
-		foreach (Shipment delivery in ShipmentList)
-		{
-			//shipOrder
-			var sentProds = delivery.Sender.ShipOrder(delivery.Products);
-			List<Product> listSent = new List<Product>();
-			foreach (var prod in sentProds)
-			{
-				listSent.Add(prod);
-			}
-			//Receive 
-			delivery.Receiver.Receive(listSent);
-		}
-		foreach (var v in SupplierList)
-		{
-			v.supplier.Process();
-		}
-		foreach (var e in EndpointList)
-		{
-			var end = (EndpointNode)e.supplier;
-			end.ProduceProduct();
-		}
-	}
-
-	public Shipment ShipmentFirstSuplier(Supplier target)
-	{
-		Shipment targetShipment = new Shipment();
-		targetShipment.Sender = target;
-		targetShipment.Products.Clear();
-		Product toShip = target.ProductInventory.First();
-		targetShipment.Products.Add(new Product());
-		targetShipment.Products[0].ProductName = toShip.ProductName;
-		targetShipment.Products[0].Quantity = 5;
-		return targetShipment;
+		ToolsForViewModel.AdvanceTimeforShipmentList(ShipmentList);
+		ToolsForViewModel.AdvancetimeForSupplierList(SupplierList);
+		ToolsForViewModel.AdvanceTimeForEndpointList(EndpointList);
 	}
 
 	public void PassTimeUntilDuration()
